FiltersPreview - removed unused code
